refactor: 重构 trait 体系，移除过时的 format_traits 模块，整合为统一的 traits 结构
